[Metadata] Remove invalid metadata manager
