+++ core/datacap-server/src/main/schema/1.17.0/schema.sql
@@ -1,0 +1,3 @@
+DELETE
+FROM `template_sql`
+WHERE `name` in ('getAllDatabase', 'getAllDatabaseAndTable', 'getAllTablesFromDatabase', 'getAllColumnsFromDatabaseAndTable');
