+++ core/datacap-web/src/views/admin/source/AdminSource.vue
@@ -67,14 +67,6 @@
                 <Button :disabled="currentUserId !== row.user.id" shape="circle" type="error" size="small" icon="md-trash"/>
               </Poptip>
             </Tooltip>
-            <Tooltip :content="$t('common.admin')" transfer>
-              <Button :disabled="currentUserId !== row.user.id || !row.available"
-                      shape="circle" type="info"
-                      size="small"
-                      icon="md-construct"
-                      :to="'/admin/source/' + row.id + '/manager'">
-              </Button>
-            </Tooltip>
             <Tooltip :content="$t('common.admin') + $t('common.beta')"
                      transfer>
               <Button :disabled="currentUserId !== row.user.id || !row.available"
@@ -82,7 +74,7 @@
                       type="info"
                       size="small"
                       icon="md-construct"
-                      :to="'/admin/source/' + row.id + '/managerBeta'">
+                      :to="'/admin/source/' + row.id + '/manager'">
               </Button>
             </Tooltip>
           </Space>
